Remove composer.lock
composer.lock is only a maintainers/contributors concern (no impact on
users installs) and we should commit it only if we put value in it.

Right now, it makes just contributors working with lowest dependencies.

I think we should better work with latest stable and let Travis test
other settings.

Also note the "require" section first-level are not even concerned as
only pdepend is controlled by composer in it.

For the record composer.lock is not committed in pdepend. If there is
no particular reason to do it differently on both projects, we should
do it the same way.

PS: Phar build fails on this PR because #623 not merged.